refactor: rename utils/props to types/props-helpers
