+++ packages/bezier-react/src/components/Center/Center.tsx
@@ -7,7 +7,7 @@ import {
   getMarginStyles,
   splitByLayoutProps,
   splitByMarginProps,
-} from '~/src/utils/props'
+} from '~/src/types/props-helpers'
 
 import { type CenterProps } from './Center.types'
