+++ packages/bezier-react/src/components/Stack/Stack.tsx
@@ -10,7 +10,7 @@ import {
   getMarginStyles,
   splitByLayoutProps,
   splitByMarginProps,
-} from '~/src/utils/props'
+} from '~/src/types/props-helpers'
 import { cssDimension } from '~/src/utils/style'
 
 import type {
